temporal-project/requirements.txt
aiohttp==3.11.18
async_lru==2.0.5
baml_py==0.89.0
fastapi==0.115.12
pymupdf>=1.24.1
Pillow==11.2.1
protobuf==6.31.0
pydantic==2.11.5
pytest==8.3.5
python-dotenv==1.1.0
qdrant_client==1.14.2
temporalio==1.11.1
typing_extensions==4.13.2
google-genai>=1.16.1
uvicorn[standard]>=0.20.0